+++ datanymizer_dumper/tests/postgres/mod.rs
@@ -1,5 +1,6 @@
 #![cfg(feature = "pg_db_tests")]
 
+mod asserts;
 mod helpers;
 
 mod connector;